Added Entity support and JPo code for testing
I copied the JPo movement code from the in-class example, and changed
him from an Image into an Entity so he can interact with platforms and
such. It turns out that the DashboardDemo code was probably not the best
sample to start from, since it's missing the Entity code and the latest
versions of a number of graphics-related classes to support it;
nonetheless, I've updated all the relevant files with their improved
counterparts in the "Gravity" sample, so they should be good now.