Made JPo jump! (sort of)
He doesn't so much "jump" as "levitate into the air when you hold
spacebar". The problem is that his velocity is being set to -100 *each
frame* that the spacebar is held down, so he's immune to gravity. A
better solution would be to only set the velocity if he's not already in
a jump.