Commented out redundant initialize() function
I already did this in Character and Raccoon (that was part of the last
commit when I made the new branch). We don't actually need these, since
all they're doing is calling the base class's initialize() with the
exact same parameters; if we leave the functions out, we'll get the
exact same behavior.