Officer increases speed if he gets too far away
Also, now that we have this, we don't need to have him bounce at the
edges of the screen any more. If you leave him in the dust and he goes
offscreen, he'll run faster and come back on screen soon enough. (Note
that the officer's "faster" speed is still slower than your speed, so if
you keep running he won't catch up, but the maps are small enough that
this isn't a big deal.)