Cleaned up more memory leaks
There still seems to be a small leak whenever you return to the menu
from the game; but this wouldn't be the same kind of leak as with the
textures, since I searched for all occurrences of "Error initializing"
and made sure they were all in initialize(). This one's small enough
that it may not be worth bothering to fix until Project 3.